refactor(app): move useCurrentRunId into resources
This is (very) widely used and is better off in a more shared location
than having all these places poke into ProtocolUpload
