+++ app/src/organisms/Devices/RobotOverviewOverflowMenu.tsx
@@ -25,7 +25,6 @@ import { Tooltip } from '../../atoms/Tooltip'
 import { ChooseProtocolSlideout } from '../../organisms/ChooseProtocolSlideout'
 import { DisconnectModal } from '../../organisms/Devices/RobotSettings/ConnectNetwork/DisconnectModal'
 import { handleUpdateBuildroot } from '../../organisms/Devices/RobotSettings/UpdateBuildroot'
-import { useCurrentRunId } from '../../organisms/ProtocolUpload/hooks'
 import { getRobotUpdateDisplayInfo } from '../../redux/robot-update'
 import { UNREACHABLE, CONNECTABLE, REACHABLE } from '../../redux/discovery'
 import { checkShellUpdate } from '../../redux/shell'
@@ -34,7 +33,7 @@ import { home, ROBOT } from '../../redux/robot-controls'
 import { useIsRobotBusy } from './hooks'
 import { useCanDisconnect } from '../../resources/networking/hooks'
 import { useIsEstopNotDisengaged } from '../../resources/devices/hooks/useIsEstopNotDisengaged'
-
+import { useCurrentRunId } from '../../resources/runs'
 import type { DiscoveredRobot } from '../../redux/discovery/types'
 import type { Dispatch, State } from '../../redux/types'
