+++ app/src/organisms/ModuleCard/hooks.tsx
@@ -11,7 +11,7 @@ import {
 import { useMostRecentCompletedAnalysis } from '../LabwarePositionCheck/useMostRecentCompletedAnalysis'
 import { MenuItem } from '../../atoms/MenuList/MenuItem'
 import { Tooltip } from '../../atoms/Tooltip'
-import { useCurrentRunId } from '../ProtocolUpload/hooks'
+import { useCurrentRunId } from '../../resources/runs'
 
 import type {
   HeaterShakerCloseLatchCreateCommand,
