+++ app/src/organisms/ProtocolUpload/hooks/index.ts
@@ -2,7 +2,6 @@ export * from './useCloseCurrentRun'
 export * from './useCurrentProtocol'
 export * from './useCurrentRun'
 export * from './useCurrentRunCommands'
-export * from './useCurrentRunId'
 export * from './useCloneRun'
 export * from './useRestartRun'
 export * from './useRunCommands'
