+++ app/src/organisms/ProtocolUpload/hooks/useCloseCurrentRun.ts
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import { useDismissCurrentRunMutation } from '@opentrons/react-api-client'
-import { useCurrentRunId } from './useCurrentRunId'
+import { useCurrentRunId } from '../../../resources/runs'
 
 import type { UseDismissCurrentRunMutationOptions } from '@opentrons/react-api-client/src/runs/useDismissCurrentRunMutation'
