+++ app/src/organisms/ProtocolUpload/hooks/useCurrentRun.ts
@@ -1,5 +1,4 @@
-import { useCurrentRunId } from './useCurrentRunId'
-import { useNotifyRunQuery } from '../../../resources/runs'
+import { useNotifyRunQuery, useCurrentRunId } from '../../../resources/runs'
 
 import type { Run } from '@opentrons/api-client'
