+++ app/src/resources/runs/index.ts
@@ -4,3 +4,4 @@ export * from './useNotifyAllRunsQuery'
 export * from './useNotifyRunQuery'
 export * from './useNotifyAllCommandsQuery'
 export * from './useNotifyAllCommandsAsPreSerializedList'
+export * from './useCurrentRunId'
